+++ Game/World.h
@@ -31,7 +31,7 @@ private:
     FrustumPlanes m_frustum;
     vec4 m_field_size;
 
-    float32 m_frame;
+    uint32 m_frame;
 
 public:
     World();
@@ -53,7 +53,7 @@ public:
     PerspectiveCamera* getBGCamera()    { return &m_camera_bg; }
     const FrustumPlanes* getViewFrustum() const { return &m_frustum; }
     const vec4& getFieldSize() const    { return m_field_size; }
-    float32 getFrame() const            { return m_frame; }
+    uint32 getFrame() const             { return m_frame; }
 
     void setFieldSize(const vec4 &v)    { m_field_size=v; }
